scrape_cdc/requirements.txt
# requirements.txt for scrape-cdc (pinned)
functions-framework==3.4.0
google-cloud-storage==2.16.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0